Port npath complexity
